moved all Register actions to new file called RegisterActions.js. Removed actions from index.js
