+++ app/src/redux/actions/index.js
@@ -1,20 +1,5 @@
 import axios from "axios";
 
-export const REGISTER_USER_LOADING = "REGISTER_USER_LOADING";
-export const REGISTER_USER_SUCCESS = "REGISTER_USER_SUCCESS";
-export const REGISTER_USER_FAILED = "REGISTER_USER_FAILED";
-
-
-export const registerUserLoading = () => ({ type: REGISTER_USER_LOADING });
-export const registerUserSuccess = data => ({
-    type: REGISTER_USER_SUCCESS,
-    payload: data
-});
-export const registerUserFailure = error => ({
-    type: REGISTER_USER_FAILED,
-    payload: error
-});
-
 export const CREATE_PROFILE_LOADING = "CREATE_PROFILE_LOADING";
 export const CREATE_PROFILE_SUCCESS = "CREATE_PROFILE_SUCCESS";
 export const CREATE_PROFILE_FAILED = "CREATE_PROFILE_FAILED";
@@ -43,19 +28,6 @@ export const sendLinkFailed = error => ({
     type: SEND_LINK_TOKEN_FAILED,
     payload: error
 });
-
-export function registerUser(data){
-    return function(dispatch) {
-        dispatch(registerUserLoading());
-        return axios.post('https://lambda-budget-blocks.herokuapp.com/api/auth/register',data)
-            .then(response=>{
-                dispatch(registerUserSuccess(response))
-            })
-            .catch(error=>{
-                dispatch(registerUserFailure(error)); 
-            })
-    }
-}
 
 
 export function sendLinkToken(token,userID){
